server/requirements.txt
Flask==3.0.0
gunicorn==20.1.0
Werkzeug==3.0.1
python-dotenv
infobip-api-python-sdk
requests
numpy
pyastar2d
opencv-python-headless
google-cloud-storage
google-cloud-vision